Refactor into a file structure that mirrors django.db.models
