+++ django_mysql/models/__init__.py
@@ -1,0 +1,5 @@
+from django_mysql.models.base import Model  # NOQA
+from django_mysql.models.query import (  # NOQA
+    ApproximateInt, SmartChunkedIterator, SmartIterator, pt_visual_explain,
+    QuerySet, QuerySetMixin
+)
