+++ django_mysql/models/base.py
@@ -1,0 +1,11 @@
+# -*- coding:utf-8 -*-
+from django.db import models
+
+from django_mysql.models.query import QuerySet
+
+
+class Model(models.Model):
+    class Meta(object):
+        abstract = True
+
+    objects = QuerySet.as_manager()
